text/documents/serializers.py
from django.contrib.auth.models import User
from documents.models import Document
from rest_framework import serializers


class UserSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = User
        fields = ('url', 'username', 'email')


class DocumentSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Document
        fields = ('id', 'title', 'owner', 'created_at', 'file', 'html', 'status')
