text/requirements.txt
Django==1.8.5
Pillow==3.0.0
amqp==1.4.7
anyjson==0.3.3
billiard==3.3.0.20
celery==3.1.18
decorator==4.0.4
djangorestframework==3.2.4
ipython==4.0.0
ipython-genutils==0.1.0
kombu==3.0.28
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.5
ptyprocess==0.5
pytz==2015.6
simplegeneric==0.8.1
traitlets==4.0.0
